requirements.txt
pandas
numpy
scikit-learn
mysql-connector-python
jieba
matplotlib
seaborn
xgboost
joblib
python-dotenv
